move reductions-related interface to pytato.reductions
